Add motor status and speed limit
Add set_motor_limits to set the speed limit when the motor is in DPS or
Position mode.
Add get_motor_status to get the motor status. Returns state (to indicate
if the motor is floating due to lower battery voltage), power (the PWM
power level), encoder position (in Degrees), and encoder speed (in
Degrees Per Second).